added stack state and settings
